backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Werkzeug==3.0.1
python-dotenv==1.0.1
cloudinary==1.39.0
httpx==0.27.0
google-cloud-aiplatform==1.43.0
google-generativeai==0.3.2
Pillow==10.2.0
PyMuPDF==1.23.26
PyPDF2==3.0.1
pdf2image==1.17.0
requests==2.31.0
gunicorn
openai
stripe==7.14.0
google-genai
markdown
psycopg2-binary